Add test for '%d' in tap name
Putting %d into the name of a tap device should make the kernel pick a
number to ensure that the resulting name is not in use. If this
mechanism is used, upon return from the TUNSETIFF ioctl, the name has to
be copied back from the if_req struct. This test verifies that all of
this happens correctly

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>