Make test_logging correctly handle case where threadname has space
Occasionally, the first logging message from a VM would be from a vcpu
thread, which is named `fc_vpu 0` or similar. The previous logic would
then split on spaces, assuming that the "tag" part of the message doesnt
contain a space.

Now, it doesn't tear apart the tag anymore if there's a space in it,
meaning we no longer get spurious failures in the above case.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>