tests: rewrite pylint test
The old test created hundreds of processes/threads (up to 1021 measured
by pgrep) to check each file one by one, and took 40s.

Worse, if it fails it only shows a single file worth of errors. There
could be more errors, but one would only find out after fixing the first
ones.

Make the directories proper modules so we can make a single pylint call,
that creates up to 15 processes/threads and takes 20s to run.

As a side bonus, we also get nice ANSI output.

Also, rename an option that is only available in pylint 2.16. We will
have to re-rename it in 2.16+.

Signed-off-by: Pablo Barbáchano <pablob@amazon.com>