fix: Slack notification on pr
This is a fix of 0d68b15d8fcfd15da70adf405d1a536c89ca3e06.
Unfortunately, the `if:` tag from GitHub action definition
doesn't work as expected when comparing with a label that
contains special characters such as `:`. I tried a couple
of ways to escape that semicolon but none of then worked.
However, filtering out the labels directly in the `run` tag
works.

Fix tested successfully in a couple of situations such as:
- adding a different tag
- adding `Status: Reviewer`
- adding a different tag (after already tagged as `Status: Reviewer`)

Signed-off-by: Trăistaru Andrei Cristian <atc@amazon.com>